+++ src/Contracts/Zoom.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace MacsiDigital\Zoom\Contracts;
+
+interface Zoom
+{
+    public function __construct($type = 'Private');
+
+    public function getClient();
+
+    public function __get($key);
+
+    public function getNode($key);
+}