+++ src/User.php
@@ -68,12 +68,11 @@ class User extends Model
 
     public function save()
     {
-        $index = $this->GetKey();
         if ($this->hasID()) {
             if (in_array('put', $this->methods)) {
-                $this->response = $this->client->patch("{$this->getEndpoint()}/{$this->id}", $this->updateAttributes);
+                $this->response = $this->client->patch("{$this->getEndpoint()}/{$this->getID()}", $this->updateAttributes);
                 if ($this->response->getStatusCode() == '200') {
-                    return $this->response->getContents();
+                    return $this;
                 } else {
                     throw new Exception($this->response->getStatusCode().' status code');
                 }
@@ -83,10 +82,9 @@ class User extends Model
                 $attributes = ['action' => 'create', 'user_info' => $this->createAttributes];
                 $this->response = $this->client->post($this->getEndpoint(), $attributes);
                 if ($this->response->getStatusCode() == '200') {
-                    $saved_item = $this->collect($this->response->getContents())->first();
-                    $this->$index = $saved_item->$index;
+                    $this->fill($this->response->getBody());
 
-                    return $this->response->getContents();
+                    return $this;
                 } else {
                     throw new Exception($this->response->getStatusCode().' status code');
                 }
@@ -97,7 +95,7 @@ class User extends Model
     public function meetings()
     {
         $meeting = new \MacsiDigital\Zoom\Meeting;
-        $meeting->setUserID($this->id);
+        $meeting->setUserID($this->getID());
 
         return $meeting;
     }
@@ -105,7 +103,7 @@ class User extends Model
     public function webinars()
     {
         $webinar = new \MacsiDigital\Zoom\Webinar;
-        $webinar->setUserID($this->id);
+        $webinar->setUserID($this->getID());
 
         return $webinar;
     }